Moved GradeForm component to its own module, imported into app
